Converts a ThinData S3 object into a SummarizedExperiment S4 object.

This only keeps the mat, design_obs, designmat, and coefmat elements of the ThinData object.

ThinDataToSummarizedExperiment(obj)

Arguments

obj
A ThinData S3 object. This is generally output by either thin_diff, thin_2group, thin_lib, thin_gene, or thin_all.

Value

A SummarizedExperiment S4 object. This is often used in Bioconductor when performing differential expression analysis.

Examples

# \donttest{
## Generate simulated data and modify using thin_diff().
## In practice, you would use real data, not simulated.
set.seed(1)
n <- 10
p <- 1000
Z <- matrix(abs(rnorm(n, sd = 4)))
alpha <- matrix(abs(rnorm(p, sd = 1)))
mat <- round(2^(alpha %*% t(Z) + abs(matrix(rnorm(n * p, sd = 5),
                                            nrow = p,
                                            ncol = n))))
design_perm <- cbind(rep(c(0, 1), length.out = n), runif(n))
coef_perm   <- matrix(rnorm(p * ncol(design_perm), sd = 6), nrow = p)
design_obs  <- matrix(rnorm(n), ncol = 1)
target_cor <- matrix(c(0.9, 0))
thout <- thin_diff(mat            = mat,
                   design_perm    = design_perm,
                   coef_perm      = coef_perm,
                   target_cor     = target_cor,
                   design_obs     = design_obs,
                   permute_method = "hungarian")

## Convert ThinData object to SummarizedExperiment object.
seobj <- ThinDataToSummarizedExperiment(thout)
class(seobj)

#> [1] "SummarizedExperiment"
#> attr(,"package")
#> [1] "SummarizedExperiment"


## The "O1" variable in the colData corresponds to design_obs.
## The "P1" and "P2" variables in colData correspond to design_perm.
seobj

#> class: SummarizedExperiment 
#> dim: 1000 10 
#> metadata(0):
#> assays(1): ''
#> rownames(1000): gene1 gene2 ... gene999 gene1000
#> rowData names(2): true_P1 true_P2
#> colnames(10): sample1 sample2 ... sample9 sample10
#> colData names(3): O1 P1 P2

# }