Estimate the effective correlation

Will return the actual correlation between the design matrix and the surrogate variables when you use permute_design.

effective_cor(design_perm, sv, target_cor, method = c("optmatch",
  "marriage"), iternum = 1000)

Arguments

design_perm
A numeric design matrix whose rows are to be permuted (thus controlling the amount by which they are correlated with the surrogate variables). The rows index the samples and the columns index the variables. The intercept should not be included (though see Section "Unestimable Components").

sv
A matrix of surrogate variables

target_cor
A numeric matrix of target correlations between the variables in design_perm and the surrogate variables. The rows index the observed covariates and the columns index the surrogate variables. The number of columns specifies the number of surrogate variables. Set this to NULL to indicate that design_perm and the surrogate variables are uncorrelated.

method
Should we use the optimal matching technique from Hansen and Klopfer (2006) ("optmatch") or the Gale-Shapley algorithm for stable marriages (marriage) (Gale and Shapley, 1962) as implemented in the matchingR package. The "optmatch" method works almost uniformly better in practice, but does take a lot more computational time if you have, say, 1000 samples.

iternum
The total number of simulated correlations to consider.